+++ app/src/main/java/com/guerra/enrico/sera/data/local/dao/TaskDao.kt
@@ -4,7 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.guerra.enrico.sera.data.local.models.Task
+import com.guerra.enrico.sera.data.models.Task
 import io.reactivex.Flowable
 
 /**
