+++ app/src/main/java/com/guerra/enrico/sera/data/local/dao/UserDao.kt
@@ -3,7 +3,7 @@ package com.guerra.enrico.sera.data.local.dao
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
-import com.guerra.enrico.sera.data.local.models.User
+import com.guerra.enrico.sera.data.models.User
 
 /**
  * Created by enrico
