+++ app/src/main/java/com/guerra/enrico/sera/data/repo/category/CategoryRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.guerra.enrico.sera.data.repo.category
 
 import com.guerra.enrico.sera.data.local.db.LocalDbManager
-import com.guerra.enrico.sera.data.local.models.Category
+import com.guerra.enrico.sera.data.models.Category
 import com.guerra.enrico.sera.data.remote.ApiError
 import com.guerra.enrico.sera.data.remote.ApiException
 import com.guerra.enrico.sera.data.remote.RemoteDataManager
@@ -10,8 +10,6 @@ import com.guerra.enrico.sera.ui.todos.CategoryFilter
 import io.reactivex.Completable
 import io.reactivex.Flowable
 import io.reactivex.Single
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
 import javax.inject.Singleton
