+++ app/src/main/java/com/guerra/enrico/sera/ui/login/LoginViewModel.kt
@@ -2,7 +2,7 @@ package com.guerra.enrico.sera.ui.login
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
-import com.guerra.enrico.sera.data.local.models.User
+import com.guerra.enrico.sera.data.models.User
 import com.guerra.enrico.sera.data.mediator.auth.GoogleSignInCallback
 import com.guerra.enrico.sera.data.result.Result
 import com.guerra.enrico.sera.ui.base.BaseViewModel
