+++ app/src/main/java/com/guerra/enrico/sera/ui/splash/SplashViewModel.kt
@@ -2,7 +2,7 @@ package com.guerra.enrico.sera.ui.splash
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
-import com.guerra.enrico.sera.data.local.models.User
+import com.guerra.enrico.sera.data.models.User
 import com.guerra.enrico.sera.data.mediator.auth.ValidateAccessToken
 import com.guerra.enrico.sera.data.result.Result
 import com.guerra.enrico.sera.ui.base.BaseViewModel
