+++ app/src/main/java/com/guerra/enrico/sera/ui/todos/add/TodoAddViewModel.kt
@@ -3,8 +3,8 @@ package com.guerra.enrico.sera.ui.todos.add
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
-import com.guerra.enrico.sera.data.local.models.Category
-import com.guerra.enrico.sera.data.local.models.Task
+import com.guerra.enrico.sera.data.models.Category
+import com.guerra.enrico.sera.data.models.Task
 import com.guerra.enrico.sera.data.mediator.category.CreateCategory
 import com.guerra.enrico.sera.data.mediator.category.LoadCategoriesFilter
 import com.guerra.enrico.sera.data.mediator.task.CreateTask
