+++ app/src/main/java/com/guerra/enrico/sera/ui/todos/filter/CategoryFilterManager.kt
@@ -1,6 +1,6 @@
 package com.guerra.enrico.sera.ui.todos.filter
 
-import com.guerra.enrico.sera.data.local.models.Category
+import com.guerra.enrico.sera.data.models.Category
 
 /**
  * Created by enrico
